server/requirements.txt
boto3==1.24.48
botocore==1.27.48
Django==4.0.6
django-cors-headers==3.13.0
django-countries==7.3.2
django-filter==22.1
django-rest-registration==0.7.2
django-storages==1.13.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.4.1
paypalrestsdk==1.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-sugar==0.9.5
python-dateutil==2.8.2
uWSGI==2.0.20
whitenoise==6.2.0
xmltodict==0.13.0
